More updates.
Made the timers configurable. The refresh time configures how
long between text redraws, which affects how quickly it scrolls.
The UI is locked to a minimum of 10 ms. You don't want to go lower
than that.

The default for polling Windows is every 1 second. This seems OK
but I made it configurable in case it causes problems.

The text looks best so far with a 15-point font. The text is OK
but not perfect. I think to get better results, I'd need to
do graphics drawing to get smooth scrolling and more control to the
user.