Version 0.77
This is a pretty substantial update.

Fixed the Windows media async interactions to only get
the global manager once for the lifetime of the plugin. This seems
to resolve some problems with the plugin running for a long time.

Changed the polling of the Windows media status to use the event
system. This eliminates one thread from the system and some complexity
coupled with the change above.

Made independent refresh timers for each context (button) so they
can all be configured and specified indepdently. I've verified this
fixes the issues with 2 buttons, both in the same and different
profiles.

Added a Debug build as part of the release so I can get info from
testers. It logs a lot of stuff, Release logs nothing.

Cleaned up the code to keep related methods together and move
uninteresting methods to the bottom of the implementation file.